requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
filelock==3.0.12
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
MouseInfo==0.1.3
numpy==1.18.1
olefile==0.46
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
pyHook==1.5.1
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.4
PyTweening==1.0.3
pywin32==227
requests==2.23.0
six==1.14.0
soupsieve==2.0.1
toml==0.10.1
tox==3.15.1
urllib3==1.25.9
virtualenv==20.0.21
wincertstore==0.2
zipp==3.1.0
